apiproject/requirements.txt
django == 3.0.3
djangorestframework == 3.11.0
psycopg2-binary == 2.8.4
requests == 2.23.0
django-cors-headers == 3.2.1
MLB-StatsAPI == 0.1.5